Almost..
I think I'm nearly there for FIFO. Yes, I know the form is pretty rough, but I just wanted it to work first. I can find the index of the container which has the most perished prodoct of interest, but I'm not sure how to make sure the product is removed from that container.  What I now have in the last line with the removing is not working (obviously), it just leads to the normal removal system.