Try in container.js
So I tried changing some things in container.js, to no avail. How does the removeItem work exactly? I did some console logging and I believe the item.values.quanity will give the quantity of the first object in the array of items, which is the same as removedQuantity, making my code redundant. But what happens when the function returns?