perishedProduct
I now am passing the 'perishedProduct' to the removeItem function, though not with much success. When there are 2 or more items of the same name in a container, it gets stuck in an infinite loop. Console logging product.values.quantity during this reveals this value will jump between the initial positive 9 (from the let's say 9 bananas in the order), to -9, to 9, -9, etc. 

I'm obviously doing something wrong, though I cant find what. Am I on the right track?