Emojis
Adds the 32 and 64 bit version of the emojis to the assets, and tries to implement the adding of said emojis to the customers. However, I'm unsure how to make each customer have its own customer-id in the html. Also for me the game still displays a ":)" after each name of the customer, though this code was removed 2 commits ago.